+++ app/(tabs)/drivers.tsx
@@ -1,11 +1,11 @@
-import { FlatList, RefreshControl, StyleSheet, ViewToken } from 'react-native';
+import DriverItem from '@/components/drivers/DriverItem';
 import NoDataFound from '@/components/NoDataFound';
 import { DriverSkeleton } from '@/components/skeleton/DriverSkeleton';
 import Header from '@/components/ui/Header';
 import { useFetchDrivers } from '@/query/hook';
 import Box from '@/theme/Box';
+import { FlatList, RefreshControl, StyleSheet, ViewToken } from 'react-native';
 import { useSharedValue } from 'react-native-reanimated';
-import DriverItem from '@/components/drivers/DriverItem';
 
 export default function HomeScreen() {
     const { data, isLoading, refetch, isRefetching } = useFetchDrivers();
@@ -29,7 +29,7 @@ export default function HomeScreen() {
                 contentContainerStyle={styles.listContainer}
                 showsVerticalScrollIndicator={false}
                 ListHeaderComponent={<Header title="F1 Drivers" backButton={false} />}
-                windowSize={1}
+                windowSize={8}
                 initialNumToRender={8}
                 maxToRenderPerBatch={8}
                 ListEmptyComponent={emptyDriver}
